requirements.txt
paho-mqtt==1.4.0
pyzmq==18.0.1
locustio==0.11.0
redis==3.3.7
